fix: fixed newline replacement to happen in backend
