+++ lib/api/ens.ts
@@ -44,7 +44,7 @@ const sanitizeOptions: sanitizeHtml.IOptions = {
     "meta",
   ],
   // URL schemes we permit
-  allowedSchemes: ["https",  "mailto", "tel"],
+  allowedSchemes: ["https", "mailto", "tel"],
   allowedSchemesByTag: {},
   allowedSchemesAppliedToAttributes: ["href", "src", "cite"],
   allowProtocolRelative: false,
@@ -69,7 +69,7 @@ export const getEnsForAddress = async (address: string | null | undefined) => {
       id: address,
       idShort,
       name: name ?? null,
-      description: sanitizeHtml(description, sanitizeOptions),
+      description: sanitizeHtml(nl2br(description), sanitizeOptions),
       url,
       twitter,
       avatar: avatar?.url ? `/api/ens-data/image/${name}` : null,
@@ -85,4 +85,16 @@ export const getEnsForAddress = async (address: string | null | undefined) => {
   };
 
   return ens;
+};
+
+export const nl2br = (str, is_xhtml = true) => {
+  if (typeof str === "undefined" || str === null) {
+    return "";
+  }
+  var breakTag =
+    is_xhtml || typeof is_xhtml === "undefined" ? "<br />" : "<br>";
+  return (str + "").replace(
+    /([^>\r\n]?)(\r\n|\n\r|\r|\n)/g,
+    "$1" + breakTag + "$2"
+  );
 };
